Add "Share By Email" extension
This is a very old request (november 2013!) to be able to share an
article directly from FreshRSS. I don't think it should be within the
core of FRSS, so I played a bit with the extension system. I keep it
very simple for now, just to try to see what was possible with the
current system. I just needed to add a new placeholder `(~ID~)` to
FreshRSS_Share.

It relies on the mailing system introduced for the "email validation"
feature.

Reference: https://github.com/FreshRSS/FreshRSS/pull/2707

Closes https://github.com/FreshRSS/FreshRSS/issues/300