Add a tool to generate the extensions.json file
Instead of manually maintaining the file, there is now a tool that grab
all repositories defined in the repositories.json file and browse through
all repositories to find extensions.
To do that, I've added a container that runs only php-cli to trigger said
tool.

The next step will be to trigger that tool automatically on a regular basis
and to update the current repository when there are changes.